# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BgmRenderSpec": {
        "additionalProperties": false,
        "properties": {
          "local_path": {
            "minLength": 1,
            "title": "Local Path",
            "type": "string"
          },
          "loop": {
            "default": true,
            "title": "Loop",
            "type": "boolean"
          },
          "volume": {
            "default": 0.12,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "local_path"
        ],
        "title": "BgmRenderSpec",
        "type": "object"
      },
      "BgmSelection": {
        "additionalProperties": false,
        "properties": {
          "local_path": {
            "minLength": 1,
            "title": "Local Path",
            "type": "string"
          },
          "loop": {
            "default": true,
            "title": "Loop",
            "type": "boolean"
          },
          "track_id": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Track Id",
            "type": "string"
          },
          "volume": {
            "default": 0.12,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "track_id",
          "local_path"
        ],
        "title": "BgmSelection",
        "type": "object"
      },
...
      "ContentCategory": {
        "enum": [
          "kindness",
          "cats_dogs",
          "forum",
          "short_video"
        ],
        "title": "ContentCategory",
        "type": "string"
      },
...
      "CreateVideoBatch": {
        "additionalProperties": false,
        "properties": {
          "bgm_loop": {
            "default": true,
            "title": "Bgm Loop",
            "type": "boolean"
          },
          "bgm_track_id": {
            "anyOf": [
              {
                "pattern": "^[a-f0-9]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bgm Track Id"
          },
          "bgm_volume": {
            "default": 0.12,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Bgm Volume",
            "type": "number"
          },
          "max_stories": {
            "default": 15,
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Max Stories",
            "type": "integer"
          },
          "story_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 15,
            "title": "Story Ids",
            "type": "array"
          },
          "subtitle": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "CreateVideoBatch",
        "type": "object"
      },
...
      "DifferentialArtReservation": {
        "additionalProperties": false,
        "properties": {
          "base_image_path": {
            "minLength": 1,
            "title": "Base Image Path",
            "type": "string"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/DifferentialArtLayer"
            },
            "maxItems": 64,
            "title": "Layers",
            "type": "array"
          }
        },
        "required": [
          "base_image_path"
        ],
        "title": "DifferentialArtReservation",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HostVisualReservations": {
        "additionalProperties": false,
        "description": "Current Remotion host contract with typed future asset reservations.\n\nThe renderer remains ``placeholder`` until a concrete Live2D or\ndifferential-art composition is installed. Assets can already be described\nwithout leaking provider-specific mappings into the orchestration service.",
        "properties": {
          "differential_art": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DifferentialArtReservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "live2d": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Live2DReservation"
              },
              {
                "type": "null"
              }
            ]
          },
          "renderer": {
            "const": "placeholder",
            "default": "placeholder",
            "title": "Renderer",
            "type": "string"
          }
        },
        "title": "HostVisualReservations",
        "type": "object"
      },
...
      "ProductionManifest": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "minLength": 1,
            "title": "Language",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "script_revision": {
            "minimum": 1.0,
            "title": "Script Revision",
            "type": "integer"
          },
          "story_id": {
            "format": "uuid",
            "title": "Story Id",
            "type": "string"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineSegment"
            },
            "minItems": 1,
            "title": "Timeline",
            "type": "array"
          },
          "total_duration_ms": {
            "exclusiveMinimum": 0.0,
            "title": "Total Duration Ms",
            "type": "integer"
          }
        },
        "required": [
          "story_id",
          "script_revision",
          "language",
          "total_duration_ms",
          "timeline"
        ],
        "title": "ProductionManifest",
        "type": "object"
      },
...
      "RemotionVideoProps": {
        "additionalProperties": false,
        "description": "Backend-owned subset of ``video/src/schema.ts``.\n\n``runtime_assets`` is intentionally absent: the local Remotion CLI stages\nfiles by content hash and owns those transient browser bindings.",
        "properties": {
          "bgm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BgmRenderSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "intro_duration_ms": {
            "default": 700,
            "maximum": 5000.0,
            "minimum": 0.0,
            "title": "Intro Duration Ms",
            "type": "integer"
          },
          "manifest": {
            "$ref": "#/components/schemas/ProductionManifest"
          },
          "subtitle": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "theme": {
            "$ref": "#/components/schemas/VideoTheme"
          },
          "title": {
            "maxLength": 240,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "transition_duration_ms": {
            "default": 180,
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Transition Duration Ms",
            "type": "integer"
          },
          "visual_reservations": {
            "$ref": "#/components/schemas/HostVisualReservations"
          }
        },
        "required": [
          "manifest",
          "title"
        ],
        "title": "RemotionVideoProps",
        "type": "object"
      },
      "RenderVideoBatch": {
        "additionalProperties": false,
        "properties": {
          "expected_batch_version": {
            "minimum": 1.0,
            "title": "Expected Batch Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_batch_version"
        ],
        "title": "RenderVideoBatch",
        "type": "object"
      },
...
      "SubmitTimelineReview": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/TimelineReviewDecision"
          },
          "expected_batch_version": {
            "minimum": 1.0,
            "title": "Expected Batch Version",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewer_id": {
            "minLength": 1,
            "title": "Reviewer Id",
            "type": "string"
          },
          "story_order": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 15,
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Order"
          }
        },
        "required": [
          "expected_batch_version",
          "decision",
          "reviewer_id"
        ],
        "title": "SubmitTimelineReview",
        "type": "object"
      },
...
      "TimelineReview": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "$ref": "#/components/schemas/TimelineReviewDecision"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "reviewed_at": {
            "format": "date-time",
            "title": "Reviewed At",
            "type": "string"
          },
          "reviewed_batch_version": {
            "minimum": 1.0,
            "title": "Reviewed Batch Version",
            "type": "integer"
          },
          "reviewer_id": {
            "minLength": 1,
            "title": "Reviewer Id",
            "type": "string"
          },
          "story_order": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 15,
            "minItems": 1,
            "title": "Story Order",
            "type": "array"
          }
        },
        "required": [
          "reviewer_id",
          "decision",
          "reviewed_batch_version",
          "story_order"
        ],
        "title": "TimelineReview",
        "type": "object"
      },
      "TimelineReviewDecision": {
        "enum": [
          "approve",
          "reject"
        ],
        "title": "TimelineReviewDecision",
        "type": "string"
      },
      "TimelineSegment": {
        "additionalProperties": false,
        "properties": {
          "audio_path": {
            "minLength": 1,
            "title": "Audio Path",
            "type": "string"
          },
          "emotion": {
            "minLength": 1,
            "title": "Emotion",
            "type": "string"
          },
          "end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "End Ms",
            "type": "integer"
          },
          "segment_id": {
            "format": "uuid",
            "title": "Segment Id",
            "type": "string"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "speaker_id": {
            "minLength": 1,
            "title": "Speaker Id",
            "type": "string"
          },
          "start_ms": {
            "minimum": 0.0,
            "title": "Start Ms",
            "type": "integer"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "visual_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Hint"
          }
        },
        "required": [
          "segment_id",
          "sequence",
          "start_ms",
          "end_ms",
          "text",
          "speaker_id",
          "emotion",
          "audio_path"
        ],
        "title": "TimelineSegment",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoBatch": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoRenderArtifact"
              },
              {
                "type": "null"
              }
            ]
          },
          "batch_id": {
            "format": "uuid",
            "title": "Batch Id",
            "type": "string"
          },
          "bgm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BgmSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "input_assets": {
            "items": {
              "$ref": "#/components/schemas/VideoInputAsset"
            },
            "maxItems": 1600,
            "minItems": 1,
            "title": "Input Assets",
            "type": "array"
          },
          "last_failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoRenderFailure"
              },
              {
                "type": "null"
              }
            ]
          },
          "remotion_props": {
            "$ref": "#/components/schemas/RemotionVideoProps"
          },
          "render_input_sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Render Input Sha256",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/VideoBatchStatus",
            "default": "PENDING_TIMELINE_REVIEW"
          },
          "stories": {
            "items": {
              "$ref": "#/components/schemas/VideoBatchStory"
            },
            "maxItems": 15,
            "minItems": 1,
            "title": "Stories",
            "type": "array"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "timeline_review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimelineReview"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "stories",
          "remotion_props",
          "input_assets",
          "render_input_sha256"
        ],
        "title": "VideoBatch",
        "type": "object"
      },
      "VideoBatchStatus": {
        "enum": [
          "PENDING_TIMELINE_REVIEW",
          "READY_TO_RENDER",
          "RENDERING",
          "RENDERED",
          "REJECTED",
          "CANCELLED",
          "FAILED"
        ],
        "title": "VideoBatchStatus",
        "type": "string"
      },
      "VideoBatchStory": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ContentCategory"
          },
          "chapter_end_ms": {
            "exclusiveMinimum": 0.0,
            "title": "Chapter End Ms",
            "type": "integer"
          },
          "chapter_start_ms": {
            "minimum": 0.0,
            "title": "Chapter Start Ms",
            "type": "integer"
          },
          "manifest": {
            "$ref": "#/components/schemas/ProductionManifest"
          },
          "reserved_at": {
            "format": "date-time",
            "title": "Reserved At",
            "type": "string"
          },
          "sequence": {
            "maximum": 14.0,
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          },
          "story_id": {
            "format": "uuid",
            "title": "Story Id",
            "type": "string"
          },
          "story_version": {
            "minimum": 1.0,
            "title": "Story Version",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Used At"
          }
        },
        "required": [
          "sequence",
          "story_id",
          "story_version",
          "category",
          "title",
          "manifest",
          "chapter_start_ms",
          "chapter_end_ms"
        ],
        "title": "VideoBatchStory",
        "type": "object"
      },
      "VideoInputAsset": {
        "additionalProperties": false,
        "description": "Immutable evidence for a local file approved for a render.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/VideoInputAssetKind"
          },
          "local_path": {
            "minLength": 1,
            "title": "Local Path",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "local_path",
          "sha256",
          "size_bytes"
        ],
        "title": "VideoInputAsset",
        "type": "object"
      },
      "VideoInputAssetKind": {
        "enum": [
          "audio",
          "bgm"
        ],
        "title": "VideoInputAssetKind",
        "type": "string"
      },
...
      "VideoRenderArtifact": {
        "additionalProperties": false,
        "properties": {
          "local_path": {
            "minLength": 1,
            "title": "Local Path",
            "type": "string"
          },
          "rendered_at": {
            "format": "date-time",
            "title": "Rendered At",
            "type": "string"
          },
          "renderer": {
            "minLength": 1,
            "title": "Renderer",
            "type": "string"
          },
          "sha256": {
            "pattern": "^[a-f0-9]{64}$",
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "local_path",
          "size_bytes",
          "sha256",
          "renderer"
        ],
        "title": "VideoRenderArtifact",
        "type": "object"
      },
      "VideoRenderFailure": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "retryable": {
            "default": true,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "title": "VideoRenderFailure",
        "type": "object"
      },
      "VideoTheme": {
        "additionalProperties": false,
        "properties": {
          "accent": {
            "default": "#85a77d",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Accent",
            "type": "string"
          },
          "background": {
            "default": "#101512",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Background",
            "type": "string"
          },
          "foreground": {
            "default": "#f3f1e8",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Foreground",
            "type": "string"
          },
          "signal": {
            "default": "#e4a853",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Signal",
            "type": "string"
          }
        },
        "title": "VideoTheme",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Review-gated multilingual content-to-audio pipeline.",
    "title": "god-news",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/video/batches": {
      "get": {
        "operationId": "listVideoBatches",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/VideoBatchStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoBatch"
                  },
                  "title": "Response Listvideobatches",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Video Batches",
        "tags": [
          "video"
        ]
      },
      "post": {
        "operationId": "createVideoBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Video Batch",
        "tags": [
          "video"
        ]
      }
    },
    "/api/v1/video/batches/{batch_id}": {
      "delete": {
        "operationId": "deleteVideoBatch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Video Batch",
        "tags": [
          "video"
        ]
      },
      "get": {
        "operationId": "getVideoBatch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Video Batch",
        "tags": [
          "video"
        ]
      }
    },
    "/api/v1/video/batches/{batch_id}/cancel": {
      "post": {
        "operationId": "cancelVideoRender",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Video Render",
        "tags": [
          "video"
        ]
      }
    },
    "/api/v1/video/batches/{batch_id}/render": {
      "post": {
        "operationId": "renderVideoBatch",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenderVideoBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render Video Batch",
        "tags": [
          "video"
        ]
      }
    },
    "/api/v1/video/batches/{batch_id}/timeline-review": {
      "post": {
        "operationId": "submitVideoTimelineReview",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTimelineReview"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoBatch"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Video Timeline Review",
        "tags": [
          "video"
        ]
      }
    },
...
  }
}